+++ tests/features/create_batch_prediction-steps.py
@@ -138,6 +138,7 @@ def i_check_batch_centroid(step, check_file):
 def i_check_batch_centroid_is_ok(step):
     assert world.api.ok(world.batch_centroid)
 
+
 @step(r'I create a batch centroid for the dataset$')
 def i_create_a_batch_prediction_with_cluster(step):
     dataset = world.dataset.get('resource')
@@ -148,3 +149,14 @@ def i_create_a_batch_prediction_with_cluster(step):
     world.location = resource['location']
     world.batch_centroid = resource['object']
     world.batch_centroids.append(resource['resource'])
+
+
+@step(r'I create a source from the batch prediction$')
+def i_create_a_source_from_batch_prediction(step):
+    batch_prediction = world.batch_prediction.get('resource')
+    resource = world.api.source_from_batch_prediction(batch_prediction)
+    world.status = resource['code']
+    assert world.status == HTTP_CREATED
+    world.location = resource['location']
+    world.source = resource['object']
+    world.sources.append(resource['resource'])